Correcciones

+++ src/main/java/com/mutant/dna/response/ResponseStats.java
@@ -1,24 +1,26 @@
 package com.mutant.dna.response;
 
+import java.math.BigDecimal;
+
 public class ResponseStats extends Response {
 
-	private Long count_mutant_dna;
-	private Long count_human_dna;
+	private BigDecimal count_mutant_dna;
+	private BigDecimal count_human_dna;
 	private String ratio;
 		
-	public Long getCount_mutant_dna() {
+	public BigDecimal getCount_mutant_dna() {
 		return count_mutant_dna;
 	}
 
-	public void setCount_mutant_dna(Long count_mutant_dna) {
+	public void setCount_mutant_dna(BigDecimal count_mutant_dna) {
 		this.count_mutant_dna = count_mutant_dna;
 	}
 
-	public Long getCount_human_dna() {
+	public BigDecimal getCount_human_dna() {
 		return count_human_dna;
 	}
 
-	public void setCount_human_dna(Long count_human_dna) {
+	public void setCount_human_dna(BigDecimal count_human_dna) {
 		this.count_human_dna = count_human_dna;
 	}
